Move globalContext out of legacy and rename legacy properties
