+++ packages/satcheljs/lib/legacy/applyMiddleware.ts
@@ -2,7 +2,7 @@ import ActionContext from './ActionContext';
 import ActionFunction from './ActionFunction';
 import DispatchFunction from './DispatchFunction';
 import Middleware from './Middleware';
-import { getGlobalContext } from './globalContext';
+import { getGlobalContext } from '../globalContext';
 
 
 export default function applyMiddleware(...middleware: Middleware[]) {
@@ -11,7 +11,7 @@ export default function applyMiddleware(...middleware: Middleware[]) {
         next = applyMiddlewareInternal(middleware[i], next);
     }
 
-    getGlobalContext().dispatchWithMiddleware = next;
+    getGlobalContext().legacyDispatchWithMiddleware = next;
 }
 
 function applyMiddlewareInternal(middleware: Middleware, next: DispatchFunction): DispatchFunction {
@@ -19,11 +19,11 @@ function applyMiddlewareInternal(middleware: Middleware, next: DispatchFunction)
 }
 
 export function dispatchWithMiddleware(action: ActionFunction, actionType: string, args: IArguments,  actionContext: ActionContext) {
-    if (!getGlobalContext().dispatchWithMiddleware) {
-        getGlobalContext().dispatchWithMiddleware = finalDispatch;
+    if (!getGlobalContext().legacyDispatchWithMiddleware) {
+        getGlobalContext().legacyDispatchWithMiddleware = finalDispatch;
     }
 
-    getGlobalContext().dispatchWithMiddleware(action, actionType, args, actionContext);
+    getGlobalContext().legacyDispatchWithMiddleware(action, actionType, args, actionContext);
 }
 
 function finalDispatch(action: ActionFunction, actionType: string, args: IArguments,  actionContext: ActionContext) {
